fix(docs): update pg.mdx
The purpose of this pull request is to improve the clarity of the enum example. I encountered some difficulties in understanding that I needed to export the custom enum in the schema file in order to generate the SQL code for creating the enum in Postgres. This change aims to provide better documentation and guidance for others who may encounter the same issue.